pyproject.toml
[tool.poetry]
name = "my_kafka"
version = "0.1.0"
description = ""
authors = ["ayekhantkhantpo <ayekhantkhantpo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pre-commit = "^2.21.0"
confluent-kafka = "^1.9.2"
kafka-python = "^2.0.2"
structlog = "^22.3.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.1"
isort = "^5.11.4"
black = "^22.12.0"
pyproject-flake8 = "^5.0.4"

[tool.poetry.scripts]
producer="my_kafka.producer:main"
consumer="my_kafka.consumer:main"
confluent_producer="my_kafka.confluent_producer:main"
confluent_consumer="my_kafka.confluent_consumer:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
